+++ precompiles/vesting/setup_test.go
@@ -1,0 +1,55 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+package vesting_test
+
+import (
+	"testing"
+
+	"github.com/cosmos/cosmos-sdk/crypto/keyring"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	"github.com/ethereum/go-ethereum/common"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	evmosapp "github.com/evmos/evmos/v14/app"
+	"github.com/evmos/evmos/v14/precompiles/vesting"
+	"github.com/evmos/evmos/v14/x/evm/statedb"
+	evmtypes "github.com/evmos/evmos/v14/x/evm/types"
+	"github.com/stretchr/testify/suite"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+var s *PrecompileTestSuite
+
+type PrecompileTestSuite struct {
+	suite.Suite
+
+	ctx        sdk.Context
+	app        *evmosapp.Evmos
+	address    common.Address
+	validators []stakingtypes.Validator
+	ethSigner  ethtypes.Signer
+	privKey    cryptotypes.PrivKey
+	signer     keyring.Signer
+	bondDenom  string
+
+	precompile *vesting.Precompile
+	stateDB    *statedb.StateDB
+
+	queryClientEVM evmtypes.QueryClient
+}
+
+func TestPrecompileTestSuite(t *testing.T) {
+	s = new(PrecompileTestSuite)
+	suite.Run(t, s)
+
+	// Run Ginkgo integration tests
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Precompile Test Suite")
+}
+
+func (s *PrecompileTestSuite) SetupTest() {
+	s.DoSetupTest()
+}