+++ scripts/protocgen.sh
@@ -21,3 +21,4 @@ done
 # move proto files to the right places
 cp -r github.com/evmos/evmos/v*/* ./
 rm -rf github.com
+